Move logging code to /common/logging and split into multiple files
