+++ extensions/ql-vscode/src/common/logging/index.ts
@@ -1,0 +1,3 @@
+export * from "./logger";
+export * from "./vscode/loggers";
+export * from "./vscode/output-channel-logger";
