+++ extensions/ql-vscode/src/common/logging/vscode/output-channel-logger.ts
@@ -1,34 +1,8 @@
 import { window as Window, OutputChannel, Progress } from "vscode";
-import { DisposableObject } from "./pure/disposable-object";
 import * as fs from "fs-extra";
 import * as path from "path";
-
-interface LogOptions {
-  /** If false, don't output a trailing newline for the log entry. Default true. */
-  trailingNewline?: boolean;
-
-  /** If specified, add this log entry to the log file at the specified location. */
-  additionalLogLocation?: string;
-}
-
-export interface Logger {
-  /** Writes the given log message, optionally followed by a newline. */
-  log(message: string, options?: LogOptions): Promise<void>;
-  /**
-   * Reveal this channel in the UI.
-   *
-   * @param preserveFocus When `true` the channel will not take focus.
-   */
-  show(preserveFocus?: boolean): void;
-
-  /**
-   * Remove the log at the specified location
-   * @param location log to remove
-   */
-  removeAdditionalLogLocation(location: string | undefined): void;
-}
-
-export type ProgressReporter = Progress<{ message: string }>;
+import { Logger, LogOptions } from "../logger";
+import { DisposableObject } from "../../../pure/disposable-object";
 
 /** A logger that writes messages to an output channel in the Output tab. */
 export class OutputChannelLogger extends DisposableObject implements Logger {
@@ -128,16 +102,4 @@ class AdditionalLogLocation {
   }
 }
 
-/** The global logger for the extension. */
-export const logger = new OutputChannelLogger("CodeQL Extension Log");
-
-/** The logger for messages from the query server. */
-export const queryServerLogger = new OutputChannelLogger("CodeQL Query Server");
-
-/** The logger for messages from the language server. */
-export const ideServerLogger = new OutputChannelLogger(
-  "CodeQL Language Server",
-);
-
-/** The logger for messages from tests. */
-export const testLogger = new OutputChannelLogger("CodeQL Tests");
+export type ProgressReporter = Progress<{ message: string }>;
