+++ extensions/ql-vscode/src/remote-queries/export-results.ts
@@ -12,7 +12,7 @@ import {
 import { Credentials } from "../authentication";
 import { UserCancellationException } from "../commandRunner";
 import { showInformationMessageWithAction } from "../helpers";
-import { logger } from "../logging";
+import { logger } from "../common";
 import { QueryHistoryManager } from "../query-history";
 import { createGist } from "./gh-api/gh-api-client";
 import { RemoteQueriesManager } from "./remote-queries-manager";
