+++ extensions/ql-vscode/src/run-queries-shared.ts
@@ -25,7 +25,8 @@ import { CodeQLCliServer } from "./cli";
 import { SELECT_QUERY_NAME } from "./contextual/locationFinder";
 import { DatabaseManager } from "./databases";
 import { DecodedBqrsChunk } from "./pure/bqrs-cli-types";
-import { logger, Logger } from "./logging";
+import { logger } from "./common";
+import { Logger } from "./common";
 import { generateSummarySymbolsFile } from "./log-insights/summary-parser";
 import { asError } from "./pure/helpers-pure";
