+++ extensions/ql-vscode/src/vscode-tests/cli-integration/new-query.test.ts
@@ -9,7 +9,7 @@ import { extensions, Uri } from "vscode";
 import { CodeQLExtensionInterface } from "../../extension";
 import { describeWithCodeQL } from "../cli";
 import { QueryServerClient } from "../../query-server/queryserver-client";
-import { logger, ProgressReporter } from "../../logging";
+import { logger, ProgressReporter } from "../../common";
 import { QueryResultType } from "../../pure/new-messages";
 import { cleanDatabases, dbLoc, storagePath } from "./global.helper";
 import { importArchiveDatabase } from "../../databaseFetcher";
