+++ extensions/ql-vscode/src/vscode-tests/cli-integration/remote-queries/variant-analysis-manager.test.ts
@@ -8,7 +8,7 @@ import {
   window,
 } from "vscode";
 import { CodeQLExtensionInterface } from "../../../extension";
-import { logger } from "../../../logging";
+import { logger } from "../../../common";
 import * as config from "../../../config";
 import {
   setRemoteControllerRepo,
